requirements.txt
notebook==7.2.1
beautifulsoup4==4.12.3
requests==2.32.3
Jinja2==3.1.4
pgvector==0.3.2
SQLAlchemy==2.0.32
python-dotenv==1.0.1
lxml
Flask==2.3.2
pillow==10.4.0
psycopg2==2.9.9
tqdm==4.66.5